# swagger/api.json
{
    "swagger": "2.0",
    "info": {
        "description": "Web simulator of the EnergieIP infrastructure",
        "termsOfService": "/tos",
        "title": "Simulator API",
        "version": "0.0.1"
    },
    "basePath": "/v1",
    "definitions": {
...
        "Error" : {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string"
                }
            }
        }
    },
    "paths": {
        "/debug/sensor/brightnessRaw": {
            "post": {
                "description": "Debug API force brightness raw detection value",
                "operationId": "debug_sensor_brightness",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Set brightness fake value on a given sensor",
                        "required": true,
                        "schema" :{
                            "type": "object",
                            "required" : [
                                "mac"
                            ],
                            "properties" :{
                                "mac": {
                                    "type" : "string",
                                    "description": "sensor mac address"
                                },
                                "brightness": {
                                    "type" : "integer",
                                    "description": "Brightness measure for the sensor in Lux"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "The Sensor",
                        "schema": {
                            "$ref": "#/definitions/Sensor"
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/Error"
                        },
                        "description": "Error detail"
                    }
                }
            }
        },
        "/debug/sensor/presence": {
            "post": {
                "description": "Debug API force presence detection value",
                "operationId": "debug_sensor_presence",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Set presence fake value on a given sensor",
                        "required": true,
                        "schema" :{
                            "type": "object",
                            "required" : [
                                "mac",
                                "presence"
                            ],
                            "properties" :{
                                "mac": {
                                    "type" : "string",
                                    "description": "sensor mac address"
                                },
                                "presence": {
                                    "type" : "boolean",
                                    "description": "Presence detection"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "The Sensor",
                        "schema": {
                            "$ref": "#/definitions/Sensor"
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/Error"
                        },
                        "description": "Error detail"
                    }
                }
            }
        },
        "/debug/sensor/temperatureRaw": {
            "post": {
                "description": "Debug API force temperature raw detection value",
                "operationId": "debug_sensor_temperature",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Set temperature fake value on a given sensor",
                        "required": true,
                        "schema" :{
                            "type": "object",
                            "required" : [
                                "mac"
                            ],
                            "properties" :{
                                "mac": {
                                    "type" : "string",
                                    "description": "sensor mac address"
                                },
                                "temperature": {
                                    "type" : "integer",
                                    "description": "Temperature measure for the sensor in 1/10 celsius"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "The Sensor",
                        "schema": {
                            "$ref": "#/definitions/Sensor"
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/Error"
                        },
                        "description": "Error detail"
                    }
                }
            }
        },
...
        "/led/switchMode": {
            "post": {
                "description": "Switch a Led driver in auto/manual mode",
                "operationId": "switch_mode_led",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The LED driver",
                        "required": true,
                        "schema" :{
                            "type": "object",
                            "required" : [
                                "mac",
                                "auto"
                            ],
                            "properties" :{
                                "mac": {
                                    "type" : "string",
                                    "description": "Driver mac address"
                                },
                                "auto": {
                                    "type" : "boolean",
                                    "description": "Driver mode: true means automatic false means manual"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Led mode switched"
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/Error"
                        },
                        "description": "Error detail"
                    }
                }
            }
        },
        "/led/brightness": {
            "post": {
                "description": "Change Led brigthness (only work in manual mode)",
                "operationId": "led_brightness",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The driver led",
                        "required": true,
                        "schema" :{
                            "type": "object",
                            "required" : [
                                "mac",
                                "setpoint"
                            ],
                            "properties" :{
                                "mac": {
                                    "type" : "string",
                                    "description": "Led mac address"
                                },
                                "setpoint": {
                                    "type" : "integer",
                                    "description": "Brightness value in percentage"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Request send to the led"
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/Error"
                        },
                        "description": "Error detail"
                    }
                }
            }
        },
...
        "/blind/switchMode": {
            "post": {
                "description": "Switch a Blind driver in auto/manual mode",
                "operationId": "switch_mode_blind",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "The driver blind",
                        "required": true,
                        "schema" :{
                            "type": "object",
                            "required" : [
                                "mac",
                                "auto"
                            ],
                            "properties" :{
                                "mac": {
                                    "type" : "string",
                                    "description": "Driver mac address"
                                },
                                "auto": {
                                    "type" : "boolean",
                                    "description": "Driver mode: true means automatic false means manual"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Blind mode switched"
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/Error"
                        },
                        "description": "Error detail"
                    }
                }
            }
        },
        "/blind/position": {
            "post": {
                "description": "Update blind position (only in manual mode)",
                "operationId": "blind_position",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Set blind position",
                        "required": true,
                        "schema" :{
                            "type": "object",
                            "required" : [
                                "mac",
                                "position"
                            ],
                            "properties" :{
                                "mac": {
                                    "type" : "string",
                                    "description": "blind mac address"
                                },
                                "position": {
                                    "type" : "integer",
                                    "description": "0: stop, 1: Openning, 2: Closing"
                                },
                                "blindNumber": {
                                    "type" : "integer",
                                    "description": "0: all, 1: only blind 1, 2: only blind2; By default: ALL"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "The Sensor",
                        "schema": {
                            "$ref": "#/definitions/Sensor"
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/Error"
                        },
                        "description": "Error detail"
                    }
                }
            }
        },
        "/blind/fin": {
            "post": {
                "description": "Update fins position (only in manual mode)",
                "operationId": "blind_fin",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Set blind position",
                        "required": true,
                        "schema" :{
                            "type": "object",
                            "required" : [
                                "mac",
                                "fin"
                            ],
                            "properties" :{
                                "mac": {
                                    "type" : "string",
                                    "description": "blind mac address"
                                },
                                "fin": {
                                    "type" : "string",
                                    "description": "Fin orientation [0,+45,-45,+90,-90]"
                                },
                                "blindNumber": {
                                    "type" : "integer",
                                    "description": "0: all, 1: only blind 1, 2: only blind2; By default: ALL"
                                }
                            }
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "The Sensor",
                        "schema": {
                            "$ref": "#/definitions/Sensor"
                        }
                    },
                    "400": {
                        "schema": {
                            "$ref": "#/definitions/Error"
                        },
                        "description": "Error detail"
                    }
                }
            }
        },
...
    }
}